move ast variants to a module
